Change data model for Skills and Characteristics
In the previous model it was very hard to implement object-json
conversions and a few features that are coming. This is why
Characteristics are now differentiated by an identifier field and Skills
are separated into definition and value. This is also closer to the
game rules structure.

Json reads and writes are also implemented.